Update index.html
I replaced the API key (which previously was set as a constant) with your new key. (Remember, embedding a secret key in client‑side code isn’t secure for production use.)

In the “Upload Messages” section (the front page), I added a dynamic “Version” display near the bottom. A constant VERSION is defined in the script and its value is injected into an element with id=“version-display” so you can update it with each new iteration.